+++ sdk/storage/azure-storage-file-datalake/azure/storage/filedatalake/aio/_path_client_async.py
@@ -48,18 +48,18 @@ class PathClient(AsyncStorageAccountHostsMixin, PathClientBase):
 
         kwargs.pop('_hosts', None)
 
-        self._blob_client = BlobClient(account_url=self._blob_account_url, container_name=file_system_name,
-                                       blob_name=path_name,
+        self._blob_client = BlobClient(account_url=self._blob_account_url, container_name=self.file_system_name,
+                                       blob_name=self.path_name,
                                        credential=credential,
                                        _hosts=self._blob_client._hosts,  # pylint: disable=protected-access
                                        **kwargs)
 
-        self._client = AzureDataLakeStorageRESTAPI(self.url, base_url=self.url, file_system=file_system_name,
-                                                   path=path_name, pipeline=self._pipeline)
+        self._client = AzureDataLakeStorageRESTAPI(self.url, base_url=self.url, file_system=self.file_system_name,
+                                                   path=self.path_name, pipeline=self._pipeline)
         self._datalake_client_for_blob_operation = AzureDataLakeStorageRESTAPI(self._blob_client.url,
                                                                                base_url=self._blob_client.url,
-                                                                               file_system=file_system_name,
-                                                                               path=path_name,
+                                                                               file_system=self.file_system_name,
+                                                                               path=self.path_name,
                                                                                pipeline=self._pipeline)
         api_version = get_api_version(kwargs)
         self._client._config.version = api_version  # pylint: disable=protected-access