+++ sdk/storage/azure-storage-file-share/tests/test_share_async.py
@@ -104,33 +104,6 @@ class StorageShareTest(AsyncStorageTestCase):
 
     @FileSharePreparer()
     @AsyncStorageTestCase.await_prepared_test
-    async def test_create_share_if_not_exists_without_existing_share(self, storage_account_name, storage_account_key):
-        self._setup(storage_account_name, storage_account_key)
-        share = self._get_share_reference()
-
-        # Act
-        created = await share.create_share_if_not_exists()
-
-        # Assert
-        self.assertTrue(created)
-        await self._delete_shares(share.share_name)
-
-    @FileSharePreparer()
-    @AsyncStorageTestCase.await_prepared_test
-    async def test_create_share_if_not_exists_with_existing_share(self, storage_account_name, storage_account_key):
-        self._setup(storage_account_name, storage_account_key)
-        share = self._get_share_reference()
-
-        # Act
-        await share.create_share()
-        created = await share.create_share_if_not_exists()
-
-        # Assert
-        self.assertIsNone(created)
-        await self._delete_shares(share.share_name)
-
-    @FileSharePreparer()
-    @AsyncStorageTestCase.await_prepared_test
     async def test_create_share_snapshot_async(self, storage_account_name, storage_account_key):
         self._setup(storage_account_name, storage_account_key)
         share = self._get_share_reference()